chore: revert projects page
